web3 constants file.
