+++ src/helpers/switchChain.ts
@@ -1,55 +1,5 @@
 import { ethers } from "ethers";
-import ethereumLogo from "assets/ethereum-logo.svg";
-import polygonLogo from "assets/UMA-OO-Polygon-tag.svg";
-
-export enum ChainId {
-  MAINNET = 1,
-  POLYGON = 137,
-}
-
-type ChainInfo = {
-  name: string;
-  chainId: ChainId;
-  logoURI: string;
-  rpcUrl?: string;
-  explorerUrl: string;
-  constructExplorerLink: (txHash: string) => string;
-  nativeCurrency: {
-    name: string;
-    symbol: string;
-    decimals: number;
-  };
-};
-
-export const CHAINS: Record<ChainId, ChainInfo> = {
-  [ChainId.MAINNET]: {
-    name: "Ethereum Mainnet",
-    chainId: ChainId.MAINNET,
-    logoURI: ethereumLogo,
-    explorerUrl: "https://etherscan.io",
-    constructExplorerLink: (txHash: string) =>
-      `https://etherscan.io/tx/${txHash}`,
-    nativeCurrency: {
-      name: "Ether",
-      symbol: "ETH",
-      decimals: 18,
-    },
-  },
-  [ChainId.POLYGON]: {
-    name: "Polygon Mainnet",
-    chainId: ChainId.POLYGON,
-    logoURI: polygonLogo,
-    rpcUrl: "https://polygon-rpc.com/",
-    explorerUrl: "https://polygonscan.com/",
-    constructExplorerLink: (txHash: string) =>
-      `https://polygonscan.com/tx/${txHash}`,
-    nativeCurrency: {
-      name: "Polygon",
-      symbol: "MATIC",
-      decimals: 18,
-    },
-  },
-};
+import { ChainId, CHAINS } from "constants/web3";
 
 export async function switchChain(
   provider: ethers.providers.JsonRpcProvider,
